Store key length as byte instead of uint32

+++ hive/lib/src/binary/binary_reader_impl.dart
@@ -220,7 +220,7 @@ class BinaryReaderImpl extends BinaryReader {
     if (keyType == FrameKeyType.uintT) {
       return readUint32();
     } else if (keyType == FrameKeyType.utf8StringT) {
-      var byteCount = readUint32();
+      var byteCount = readByte();
       return BinaryReader.utf8Decoder.convert(viewBytes(byteCount));
     } else {
       throw HiveError('Unsupported key type. Frame might be corrupted.');
